[PATCH] For OE "morty", use IMGDEPLOYDIR for deployment
Starting with the morty release of OpenEmbedded, the image creation is
also under sstate control. This introduced the IMGDEPLOYDIR variable
for intermediate storage. To be compatible with both older and newer
OE versions, set a default IMGDEPLOYDIR if not set already, and use
that in the image creation script where appropriate.
( http://forums.openpli.org/topic/40129-openpli-5-still-next-master/?view=findpost&p=624721 )