h3: move plugin vfdcontrol from image to feed
In this receiver is used skin_text.xml in which using converter VfdDisplay can introduce the same functionality as in the plugin.
Therefore it is not necessary to use the plugin together with skin_text.xml which is used in any case.
If anyone still wants to use this plugin, he can be installed from the feed.